requirements.txt
flask
flask-cors

flask-sqlalchemy
pymysql
flask-migrate

flask-caching
redis

pyjwt
werkzeug

python-dotenv
gunicorn

requests
chardet

apscheduler
beautifulsoup4
lxml
